Updated all change notifications to use changeSubject approach.

+++ kohese-portal/common/models/item-proxy.js
@@ -117,6 +117,15 @@ class ItemProxy {
 
     proxy.calculateTreeHash();
 
+    if(!tree.loading){
+      tree.changeSubject.next({
+        type: 'create',
+        kind: proxy.kind,
+        id: proxy.item.id,
+        proxy: proxy
+      });
+    }
+
     return proxy;
   }
 
@@ -130,9 +139,14 @@ class ItemProxy {
 
     tree.loading = true;
 
+    tree.changeSubject.next({
+      type: 'loading'
+    });
+
     rootProxy.visitChildren(null, null, (childProxy) => {
       childProxy.deleteItem();
     });
+    
   }
   
   //////////////////////////////////////////////////////////////////////////
@@ -156,6 +170,9 @@ class ItemProxy {
   static loadingComplete() {
     tree.loading = false;
     ItemProxy.calculateAllTreeHashes();
+    tree.changeSubject.next({
+      type: 'loaded'
+    });
   }
   
   //////////////////////////////////////////////////////////////////////////
@@ -945,7 +962,7 @@ class ItemProxy {
   //
   //////////////////////////////////////////////////////////////////////////
   updateItem(modelKind, withItem) {
-    console.log('!!! Updating ' + modelKind + ' - ' + this.item.id);
+//    console.log('!!! Updating ' + modelKind + ' - ' + this.item.id);
     
     var validation = ItemProxy.validateItemContent(modelKind, withItem);
     
@@ -1018,6 +1035,16 @@ class ItemProxy {
         // delete the analysis in case some of the requisite data was updated
         delete this.analysis;
     }
+    
+    if(!tree.loading){
+      tree.changeSubject.next({
+        type: 'update',
+        kind: this.kind,
+        id: this.item.id,
+        proxy: this
+      });
+    }
+
   }
 
   //////////////////////////////////////////////////////////////////////////
@@ -1046,12 +1073,14 @@ class ItemProxy {
 //        console.log('::: -> Not removing ' + this.item.name);        
       } else {
 //        console.log('::: -> Removing all references');
-        tree.changeSubject.next({
-          type: 'delete',
-          kind: this.kind,
-          id: this.item.id,
-          proxy: this
-        });
+        if(!tree.loading){
+          tree.changeSubject.next({
+            type: 'delete',
+            kind: this.kind,
+            id: this.item.id,
+            proxy: this
+          });          
+        }
         delete tree.proxyMap[byId];
       }
     } else {
@@ -1059,12 +1088,14 @@ class ItemProxy {
       if (this.children.length !== 0) {
         if (!attemptToDeleteRestrictedNode){
 //          console.log('::: -> Node still has children');
-          tree.changeSubject.next({
-            type: 'delete',
-            kind: this.kind,
-            id: this.item.id,
-            proxy: this
-          });
+          if(!tree.loading){
+            tree.changeSubject.next({
+              type: 'delete',
+              kind: this.kind,
+              id: this.item.id,
+              proxy: this
+            });
+          }
           createMissingProxy(byId);          
         }
       } else {
@@ -1072,12 +1103,14 @@ class ItemProxy {
 //          console.log('::: -> Not removing ' + this.item.name);                  
         } else {
 //          console.log('::: -> Removing all references');
-          tree.changeSubject.next({
-            type: 'delete',
-            kind: this.kind,
-            id: this.item.id,
-            proxy: this
-          });
+          if(!tree.loading){
+            tree.changeSubject.next({
+              type: 'delete',
+              kind: this.kind,
+              id: this.item.id,
+              proxy: this
+            });
+          }
           delete tree.proxyMap[byId];
         }
       }
